+++ src/bsense_experiment/__init__.py
@@ -1,4 +1,3 @@
 """BSense-R LSL experiment control package."""
 
-__version__ = "0.1.1"
-
+__version__ = "0.2.5"